# docs/industrial/openapi.json
{
  "components": {
    "schemas": {
...
      "Citation": {
        "properties": {
          "chunk_id": {
            "format": "uuid",
            "title": "Chunk Id",
            "type": "string"
          },
          "quote": {
            "maxLength": 900,
            "minLength": 1,
            "title": "Quote",
            "type": "string"
          }
        },
        "required": [
          "chunk_id",
          "quote"
        ],
        "title": "Citation",
        "type": "object"
      },
      "Claim": {
        "properties": {
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "title": "Citations",
            "type": "array"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "citations"
        ],
        "title": "Claim",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IndexAccepted": {
        "properties": {
          "generation_id": {
            "format": "uuid",
            "title": "Generation Id",
            "type": "string"
          },
          "job_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "generation_id",
          "job_id",
          "state"
        ],
        "title": "IndexAccepted",
        "type": "object"
      },
      "IndexRequest": {
        "additionalProperties": false,
        "properties": {
          "force_rebuild": {
            "default": false,
            "title": "Force Rebuild",
            "type": "boolean"
          },
          "model_profile_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Profile Id"
          },
          "retrieval_profile": {
            "const": "default-v1",
            "default": "default-v1",
            "title": "Retrieval Profile",
            "type": "string"
          }
        },
        "title": "IndexRequest",
        "type": "object"
      },
      "IndexView": {
        "properties": {
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "embedded_count": {
            "title": "Embedded Count",
            "type": "integer"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kb_id": {
            "format": "uuid",
            "title": "Kb Id",
            "type": "string"
          },
          "model_digest": {
            "title": "Model Digest",
            "type": "string"
          },
          "model_profile_id": {
            "format": "uuid",
            "title": "Model Profile Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kb_id",
          "status",
          "chunk_count",
          "embedded_count",
          "model_profile_id",
          "model_digest",
          "error_code"
        ],
        "title": "IndexView",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "QueryRequest": {
        "additionalProperties": false,
        "properties": {
          "conversation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "knowledge_base_id": {
            "format": "uuid",
            "title": "Knowledge Base Id",
            "type": "string"
          },
          "method": {
            "default": "hybrid",
            "enum": [
              "bm25",
              "dense",
              "hybrid"
            ],
            "title": "Method",
            "type": "string"
          },
          "profile": {
            "default": "evidence",
            "enum": [
              "evidence",
              "ollama"
            ],
            "title": "Profile",
            "type": "string"
          },
          "question": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "knowledge_base_id",
          "question"
        ],
        "title": "QueryRequest",
        "type": "object"
      },
      "QueryResponse": {
        "properties": {
          "claims": {
            "items": {
              "$ref": "#/components/schemas/Claim"
            },
            "title": "Claims",
            "type": "array"
          },
          "conversation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "effective_question": {
            "title": "Effective Question",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "generation_assessment": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Assessment"
          },
          "query_id": {
            "format": "uuid",
            "title": "Query Id",
            "type": "string"
          },
          "query_job_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Job Id"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "scope": {
            "$ref": "#/components/schemas/QueryScope"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "title": "Sources",
            "type": "array"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "answered",
              "evidence_found",
              "no_evidence",
              "needs_scope",
              "clarify",
              "partial",
              "failed",
              "cancelled",
              "superseded"
            ],
            "title": "Status",
            "type": "string"
          },
          "timing": {
            "$ref": "#/components/schemas/Timing"
          },
          "warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warning"
          }
        },
        "required": [
          "query_id",
          "status",
          "scope",
          "question",
          "effective_question"
        ],
        "title": "QueryResponse",
        "type": "object"
      },
      "QueryScope": {
        "properties": {
          "data_epoch": {
            "title": "Data Epoch",
            "type": "integer"
          },
          "index_generation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Index Generation Id"
          },
          "kb_id": {
            "format": "uuid",
            "title": "Kb Id",
            "type": "string"
          }
        },
        "required": [
          "kb_id",
          "data_epoch",
          "index_generation_id"
        ],
        "title": "QueryScope",
        "type": "object"
      },
...
      "Source": {
        "properties": {
          "collection": {
            "title": "Collection",
            "type": "string"
          },
          "context_origin": {
            "default": "retrieved",
            "title": "Context Origin",
            "type": "string"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "end_line": {
            "title": "End Line",
            "type": "integer"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "heading": {
            "title": "Heading",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "parse_run_id": {
            "format": "uuid",
            "title": "Parse Run Id",
            "type": "string"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "revision_id": {
            "format": "uuid",
            "title": "Revision Id",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "start_line": {
            "title": "Start Line",
            "type": "integer"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "document_id",
          "revision_id",
          "parse_run_id",
          "filename",
          "collection",
          "version",
          "heading",
          "text",
          "page",
          "start_line",
          "end_line",
          "score",
          "rank"
        ],
        "title": "Source",
        "type": "object"
      },
...
      "Timing": {
        "properties": {
          "generation_ms": {
            "default": 0,
            "title": "Generation Ms",
            "type": "number"
          },
          "queue_ms": {
            "default": 0,
            "title": "Queue Ms",
            "type": "number"
          },
          "retrieval_ms": {
            "default": 0,
            "title": "Retrieval Ms",
            "type": "number"
          },
          "total_ms": {
            "default": 0,
            "title": "Total Ms",
            "type": "number"
          }
        },
        "title": "Timing",
        "type": "object"
      },
      "TraceView": {
        "properties": {
          "details": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Details",
            "type": "object"
          },
          "query_id": {
            "format": "uuid",
            "title": "Query Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "query_id",
          "status",
          "details"
        ],
        "title": "TraceView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "in": "cookie",
        "name": "__Host-tracedesk-session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "TraceDesk Team API",
    "version": "0.3.0-dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/index": {
      "post": {
        "operationId": "upgrade_api_index_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Upgrade"
      }
    },
...
    "/api/v1/index-generations/{generation_id}": {
      "get": {
        "operationId": "get_api_v1_index_generations__generation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "generation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Generation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Get"
      }
    },
    "/api/v1/index-generations/{generation_id}/activate": {
      "post": {
        "operationId": "activate_api_v1_index_generations__generation_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "generation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Generation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Activate"
      }
    },
    "/api/v1/index-generations/{generation_id}/rollback": {
      "post": {
        "operationId": "rollback_api_v1_index_generations__generation_id__rollback_post",
        "parameters": [
          {
            "in": "path",
            "name": "generation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Generation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Rollback"
      }
    },
...
    "/api/v1/knowledge-bases/{kb_id}/index-generations": {
      "post": {
        "operationId": "create_api_v1_knowledge_bases__kb_id__index_generations_post",
        "parameters": [
          {
            "in": "path",
            "name": "kb_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Kb Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndexRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Create"
      }
    },
...
    "/api/v1/queries": {
      "post": {
        "operationId": "query_api_v1_queries_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Query"
      }
    },
    "/api/v1/queries/{query_id}": {
      "get": {
        "operationId": "result_api_v1_queries__query_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "query_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Query Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Result"
      }
    },
    "/api/v1/queries/{query_id}/export.md": {
      "get": {
        "operationId": "export_api_v1_queries__query_id__export_md_get",
        "parameters": [
          {
            "in": "path",
            "name": "query_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Query Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Export"
      }
    },
    "/api/v1/queries/{query_id}/trace": {
      "get": {
        "operationId": "trace_api_v1_queries__query_id__trace_get",
        "parameters": [
          {
            "in": "path",
            "name": "query_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Query Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Trace"
      }
    },
...
  }
}